CHANGE INDEX TYPES Ix, Ixs TO usize, isize
This is important for iterator efficiency (outside of the already
efficient contiguous cases). Counted and indexed loops compile much
better using machine-size indices, so for example strided inner
axes are iterated much quicker.

It also allows bigger arrays (bigger axis dimension and bigger strides).

Arrays and views become larger in size, but the difference should be
small. In the bigger data use cases, this difference should be
insignificant and the iteration wins significant.

[breaking-change]