FEAT: Update to newest (c)blas-sys, blas-src
Update dependencies, code and documentation for new changes to blas-sys,
cblas-sys and blas-src.

ndarray now links to blas-src. Use it to configure the blas backend for
ndarray.

Also add a subcrate that tests this setup, so that it works. It's using
system openblas, which we have set up so that it should work in travis.