FEAT: Add .reborrow() methods to array views
`ArrayView<'a, T, D>` is invariant in parameters `T` and `'a`, because
`ArrayBase<S, D>` is invariant in the parameter `S` because the array
base is using the projection `<S as Data>::Elem` in the struct
definition.

This means that the equivalent of the `&'a T -> &'b T` conversion (valid
whenever `'a: 'b` is not naturally available for array views; we can
provide it manually, with a method.