Adjust constraints on ddof for var_axis and std_axis
Requiring `ddof` to not be < 0 allows us to ignore weird edge cases.
Allowing `ddof` to be equal to the length of the axis allows
computation of the population variance (`ddof` = 0) for a zero-length
axis and computation of the sample variance (`ddof` = 1) for an axis
of length one. (The result in these cases is an array of NaNs.)

Note that this is a breaking change because of the new constraint that
`ddof` must not be less than zero.