Switch var_axis and std_axis to use FromPrimitive
The documentation for the `Zero` and `One` traits says only that they
are the additive and multiplicative identities; it doesn't say
anything about converting an integer to a float by adding `One::one()`
to `Zero::zero()` repeatedly. Additionally, it's nice to panic early
instead of waiting until after the sum has been calculated.