FIX: .axis_chunks_iter for zero stride and zero sized types
Two special cases were identified, first by @jturner314, that break the
logic that picks out the last and uneven chunk in axis_chunks_iter/_mut;
we can't rely on a specific pointer value to recognize this chunk, as
the cases in the new test show.

Instead use the indices and length stored in the iterator.

Fixes #542