DOC: Use box drawing art for illustrating an example for ArrayView::split_at
This is a proof of concept.

Box drawings are relatively easy to produce, easy to distribute and
maintain in all places rustdoc renders, readable in plain text and
when rendered, and they have a consistent style.

The biggest positive is that it enables us to produce many
illustrations.

They have the drawback that they restrict us to mostly 2-D (but that makes
for easier to understand illustrations anyway), and it doesn't look very
impressive.