Fix try_ensure_unique for CowRepr
The old implementation incorrectly handled the strides when the array
was not in standard layout. (It created a `Vec` corresponding to an
array in standard layout but didn't update the strides accordingly.)
The new implementation calls `.to_owned()`, which handles the problem
of efficiently creating an owned array, and then moves all the field
values. Additionally, the `ensure_is_owned` function has been removed
because it's equivalent to `DataMut::ensure_unique`.