Impl clone_from_with_ptr for differing CowRepr variants
The old implementation panicked when the variants were different; the
new implementation performs the necessary clone instead of panicking.
The variant of `self` after calling `clone_from_with_ptr` should match
the variant of `other`. If the variants are initially different, the
data from `other` needs to be cloned, and `self` needs to be changed
to that variant.