Avoid re-borrowing in multislice! macro
The old implementation always called `.view_mut()` on the input array,
which meant that it was not possible to use `multislice!` to split an
`ArrayViewMut` instance into pieces with the same lifetime as the
original view. Now, `multislice!` uses `ArrayViewMut::from()` instead
so that this works. The primary disadvantage is that an explicit
borrow is necessary in many cases, which is less concise.

Closes #687.