Use NonNull::offset method and factor out nonnull constructors
Add two constructors that suffice for ndarray:

- Get NonNull<T> from Vec<T>'s buffer pointer
- Get NonNull<T> from raw pointer, and check if it's nonnull with
a debug assertion

In all other cases we can use .offset() on the NonNull (method from
rawpointer 0.2, so we don't have to explicit roundtrip through raw
pointers when we offset NonNull<T>.

For safety, see this documentation in the rawpointer crate, which
says why using .offset() on NonNull is safe.

NonNull<T> supports the same offsetting methods under the same safety
constraints as the other raw pointer implementations.

There is no difference - both when offsetting *mut T and NonNull<T>, the
offset is only well defined if we remain inside the same object or
one-past the end, and we can never land in a null pointer while obeying
those rules.