FIX: Reduce genericity of inner format_array function
This is a slight de-bloating of this function, by converting to a dyn
dimensioned array before entering the main code. This saves a lot of code size,
even though there is more than can be done.

This also allows setting the top level parameters in only one place
(starting at depth 0 etc).