FIX: Remove itertools as a dependency
We use `izip!()` as the only major feature from itertools, and can
remove it to save build time and two small crates as deps.

`izip!()` is license compatible and originally written by @krdln
(Michał Krasnoborski) and @bluss in itertools, so we copy it into the
crate.

The free functions zip(i, j) and enumerate(i) are just style
preferences, written by me, and simple to copy.

Itertools is still used as a dev-dependency.