FEAT: Add methods RawArrayView/Mut::cast
Add methods for raw view casts, this makes it easier to build upon the
raw array view functionality.

Of course, adding more capabilities for the unsafe APIs opens for users
to make mistakes, but we need these capabilities ourselves for
developing features for ndarray.

The reason for this to exist is for the `ArrayViewMut<T>` →
`ArrayView<Cell<T>>` transformation demonstrated in the tests. However, we
don't add a method for this directly, at least not yet. The reason is
that we'd like to have a type that is exactly like `Cell<T>` but also
supports the arithmetic ops.