Issue/655: Standard deviation and variance
This PR implements std and var methods using the same
method as used for var_axis, std_axis.
The variance is computed by flattened the array into
a 1D array of the .len() of original array.

The return type is scalar A instead of Array<A, D:Smaller>

An attempt was made to refactor var_axis to use var, but this
resulted in a performance regression so original implementations
are kept and we accept some code duplication.

```
    pub fn var_axis(&self, axis: Axis, ddof: A) -> Array<A, D::Smaller>
    where
        A: Float + FromPrimitive,
        D: RemoveAxis,
    {
        let mut sum_sq = Array::<A, D::Smaller>::zeros(self.dim.remove_axis(axis));

        for (lane, sum) in self.lanes(axis).into_iter().zip(sum_sq.iter_mut()) {
            *sum = lane.var(ddof);
        }
        sum_sq
    }
```