FEAT: New method .assume_init() for all array types
Add representation trait RawDataSubst that array storage types correctly
(emulation of higher kinded types). Implement it as a transmute-like
operation, using the representation guarantees that MaybeUninit
documents. Make it a public method - avaialable on all array types -
owned arrays, arcarray, array views and so on.

The safety percussions of exposing this transformation are only explored
to an extent, but it's a type change without value conversion so there
should not be much that can go wrong.

Interesting aspects include being able to have one ArcArray handle with
the old type and one with the new type, or one owned array with old and
a view with new.