FEAT: Keep track of "layout tendency" in Zip for better performance
Support both unroll over c- and f-layout preferred axis in Zip inner loop
(the fallback when inputs are not all contiguous and same layout).

Keep a tendency score when building the Zip, so that we know if the
inputs are tending to be c- or f- layout.

This improves performance on the just added zip_indexed_ff benchmark, so
that it seems to match its (already fast) cc counterpart.