FEAT: Let 1D array views be IndexedParallelIterators
Impl this trait for 1D parallel iterators.

One piece is missing - opt_len - but it's mostly used for .collect()
optimization. We forgo this for now, because it would require some type
system delusion to conditionally call the right drive method inside
Parallel<ArrayView<..>>::drive_unindexed; conditional on if D is Ix1 or
not.