FIX: Factor out the inner loops in Zip apply_core_*
The innermost loop is the stretch of constant stride (for the contiguous
case, this is the whole loop); with a lot of function arguments, this
part is actually commmon to the three loops (contig, c, and f).

Also make a small simplification of this inner loop; use a while loop to
have less code to compile in this much instantiated function.