API: Rename .genrows/gencolumns/_mut into .rows/columns/_mut
Rename these to give them clearer names. Rows is for rows (and
generalized rows, which is a bit strange). Columns is for columns (and
generalized ones).

Generalized just means that we take the concept of rows/columns and
somehow extend it to n-dimension arrays.

The old names .genrows/gencolums/_mut are deprecated