API: Remove deprecated ArrayView::into_slice (use to_slice)
This could really use either name - `into_slice` for the "into" semantic
of preserving the lifetime parameter, or the `to_slice` name which is
customary for `&self` methods.

Both names have the problem that they compete with
`ArrayBase::as_slice()`, but the doc has been improved to explain some
of that.