MAINT: Fix rayon, serde, blas features for no-std changes
rayon: implies std, because threads to
serde: works fine without std
blas: works fine without std if we use libc (libc is a new dependency
for this feature flag, but it's already used by blas transitively).

A slight hack is used for rayon (renaming the dependency) because we
need to use a feature flag to enable std in this crate too.