Make SliceInfo be Sized, and remove pointer casts
The original reason for `SliceInfo` being `?Sized` and these
conversions was to work around the limitations of the slicing methods
taking a type involving `Dimension::SliceArg`. Now that the slicing
methods take `I: SliceArg<D>` as the parameter type, these conversions
are no longer necessary, and `SliceInfo` doesn't need to be `?Sized`.