FIX: For stride == 1 axes, require len > 1 to determine C/F preference
The issue was that otherwise, this array:
shape: [1, 256]
strides: [1, 256]

Which was "correctly" determined to be F-preference because it's
sliced out of an f-order array. In reality it is discontiguous; every
element is 256 steps from the next. This should have layout None, the
same as this array would have (ignoring 1-len axes).

shape: [256]
strides: [256]