Update approx to >=0.4, <0.6 to support approx 0.5
`approx` 0.5 updates its `num-complex` dependency to 0.4, which is the
same version of `num-complex` which `ndarray` depends on. So, it's
very useful for `ndarray` to support `approx` 0.5. Instead of
expanding the range of the `approx` versions, another option would be
to add another feature for `approx` 0.5, but I don't think the
additional complexity would be worth it. When we release the next
breaking release of `ndarray`, I think we should change this version
specification to "0.5".