Add Parallel::with_min_len to limit splitting
Splitting of parallel iterators working `ArrayView` and `Zip` is currently
not limited and continues until only one element is left which can lead to
excessive overhead for algorithms formulated in terms of these iterators.

Since these iterators are also not indexed, Rayon's generic
`IndexedParallelIterator::with_min_len` does not apply. However, since the
number of elements is known and currently checked against to one to determine if
another split is possible, it appears straight-forward to replace this constant
by a parameter and make it available to the user via a `Parallel::with_min_len`
inherent method.