Fix comparison of slicing with negative step in `ndarray` with that in NumPy
The existing text implies that `ndarray` and NumPy behave the same way when `step = -1`.
However this is not the case. The behavior of NumPy is illustrated below:

In [1]: import numpy as np
In [2]: x = np.arange(10)
In [3]: print(x[1:5:-1])
[]
In [4]: print(x[5:1:-1])
[5 4 3 2]

The analogous results in `ndarray` would be `array![4, 3, 2, 1]` and `array![]`, respectively.