Check for aliasing in RawViewMut::from_shape_ptr
It should deny overlapping indexes. Because it's an unsafe function and
we only do these checks on best-effort basis (caller has to ensure they
are correct), it's a debug assertion.

Removed the small addition to a doc comment that was done in a previous
PR (it was an unfinished idea).

Similar to pr #1410
Finishes and closes #919